# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Web3 Job Marketplace API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type CategoryType = (typeof CategoryType)[keyof typeof CategoryType];

export const CategoryType = {
  manual: "manual",
  online: "online",
} as const;

export interface Category {
  id: number;
  name: string;
  type: CategoryType;
  industry: string;
  payLevel?: string | null;
  createdAt: string;
}

export type IndustryType = (typeof IndustryType)[keyof typeof IndustryType];

export const IndustryType = {
  manual: "manual",
  online: "online",
} as const;

export interface Industry {
  industry: string;
  type: IndustryType;
  count: number;
}

export type WorkerCategoryType =
  (typeof WorkerCategoryType)[keyof typeof WorkerCategoryType];

export const WorkerCategoryType = {
  manual: "manual",
  online: "online",
} as const;

export type WorkerRateType =
  (typeof WorkerRateType)[keyof typeof WorkerRateType];

export const WorkerRateType = {
  hourly: "hourly",
  task: "task",
} as const;

export interface Worker {
  id: number;
  walletAddress: string;
  displayName: string;
  bio?: string | null;
  avatarUrl?: string | null;
  categoryId: number;
  categoryName: string;
  categoryType: WorkerCategoryType;
  industry: string;
  rateAmount: string;
  rateCurrency: string;
  rateType: WorkerRateType;
  country?: string | null;
  state?: string | null;
  city?: string | null;
  /** Native wallet address for BTC/SOL payments (optional for ETH/MATIC/BNB/USDC/USDT workers) */
  paymentAddress?: string | null;
  /**
   * @minimum 5
   * @maximum 20
   */
  advanceDepositPercent?: number | null;
  isAvailable: boolean;
  completedJobs: number;
  rating?: number | null;
  createdAt: string;
}

export type CreateWorkerBodyRateCurrency =
  (typeof CreateWorkerBodyRateCurrency)[keyof typeof CreateWorkerBodyRateCurrency];

export const CreateWorkerBodyRateCurrency = {
  ETH: "ETH",
  MATIC: "MATIC",
  USDC: "USDC",
  USDT: "USDT",
  BNB: "BNB",
  BTC: "BTC",
  SOL: "SOL",
} as const;

export type CreateWorkerBodyRateType =
  (typeof CreateWorkerBodyRateType)[keyof typeof CreateWorkerBodyRateType];

export const CreateWorkerBodyRateType = {
  hourly: "hourly",
  task: "task",
} as const;

export interface CreateWorkerBody {
  walletAddress: string;
  displayName: string;
  bio?: string | null;
  avatarUrl?: string | null;
  categoryId: number;
  rateAmount: string;
  rateCurrency: CreateWorkerBodyRateCurrency;
  rateType: CreateWorkerBodyRateType;
  country?: string | null;
  state?: string | null;
  city?: string | null;
  /** Native wallet address for BTC/SOL payments */
  paymentAddress?: string | null;
  /**
   * @minimum 5
   * @maximum 20
   */
  advanceDepositPercent?: number | null;
}

export type UpdateWorkerBodyRateCurrency =
  (typeof UpdateWorkerBodyRateCurrency)[keyof typeof UpdateWorkerBodyRateCurrency];

export const UpdateWorkerBodyRateCurrency = {
  ETH: "ETH",
  MATIC: "MATIC",
  USDC: "USDC",
  USDT: "USDT",
  BNB: "BNB",
  BTC: "BTC",
  SOL: "SOL",
} as const;

export type UpdateWorkerBodyRateType =
  (typeof UpdateWorkerBodyRateType)[keyof typeof UpdateWorkerBodyRateType];

export const UpdateWorkerBodyRateType = {
  hourly: "hourly",
  task: "task",
} as const;

export interface UpdateWorkerBody {
  displayName?: string;
  bio?: string | null;
  avatarUrl?: string | null;
  rateAmount?: string;
  rateCurrency?: UpdateWorkerBodyRateCurrency;
  rateType?: UpdateWorkerBodyRateType;
  country?: string | null;
  state?: string | null;
  city?: string | null;
  /** Native wallet address for BTC/SOL payments */
  paymentAddress?: string | null;
  /**
   * @minimum 5
   * @maximum 20
   */
  advanceDepositPercent?: number | null;
  isAvailable?: boolean;
}

export type BookingStatus = (typeof BookingStatus)[keyof typeof BookingStatus];

export const BookingStatus = {
  pending: "pending",
  confirmed: "confirmed",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface Booking {
  id: number;
  workerId: number;
  workerName: string;
  workerWallet: string;
  clientWallet: string;
  categoryName: string;
  description?: string | null;
  rateAmount: string;
  rateCurrency: string;
  status: BookingStatus;
  txHash?: string | null;
  scheduledAt?: string | null;
  /**
   * @minimum 1
   * @maximum 5
   */
  clientRating?: number | null;
  clientReview?: string | null;
  createdAt: string;
}

export interface CreateBookingBody {
  workerId: number;
  clientWallet: string;
  description?: string | null;
  scheduledAt?: string | null;
}

export type UpdateBookingBodyStatus =
  (typeof UpdateBookingBodyStatus)[keyof typeof UpdateBookingBodyStatus];

export const UpdateBookingBodyStatus = {
  pending: "pending",
  confirmed: "confirmed",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface UpdateBookingBody {
  status?: UpdateBookingBodyStatus;
  txHash?: string | null;
  /**
   * @minimum 1
   * @maximum 5
   */
  clientRating?: number | null;
  clientReview?: string | null;
}

export interface MarketplaceStats {
  totalWorkers: number;
  availableWorkers: number;
  totalBookings: number;
  completedBookings: number;
  totalCategories: number;
  manualCategories: number;
  onlineCategories: number;
}

export type CategoryStatType =
  (typeof CategoryStatType)[keyof typeof CategoryStatType];

export const CategoryStatType = {
  manual: "manual",
  online: "online",
} as const;

export interface CategoryStat {
  categoryId: number;
  categoryName: string;
  industry: string;
  type: CategoryStatType;
  workerCount: number;
}

export interface Message {
  id: number;
  fromWallet: string;
  toWallet: string;
  content: string;
  isRead: boolean;
  createdAt: string;
}

export interface Conversation {
  otherWallet: string;
  lastMessage: string;
  lastMessageAt: string;
  unreadCount: number;
}

export interface SendMessageBody {
  fromWallet: string;
  toWallet: string;
  content: string;
}

export type ListCategoriesParams = {
  type?: ListCategoriesType;
  industry?: string;
  search?: string;
};

export type ListCategoriesType =
  (typeof ListCategoriesType)[keyof typeof ListCategoriesType];

export const ListCategoriesType = {
  manual: "manual",
  online: "online",
} as const;

export type ListWorkersParams = {
  type?: ListWorkersType;
  categoryId?: number;
  search?: string;
  country?: string;
  state?: string;
  city?: string;
  minRate?: number;
  maxRate?: number;
  available?: boolean;
  /**
   * @minimum 1
   * @maximum 5
   */
  minRating?: number;
  sortBy?: ListWorkersSortBy;
};

export type ListWorkersType =
  (typeof ListWorkersType)[keyof typeof ListWorkersType];

export const ListWorkersType = {
  manual: "manual",
  online: "online",
} as const;

export type ListWorkersSortBy =
  (typeof ListWorkersSortBy)[keyof typeof ListWorkersSortBy];

export const ListWorkersSortBy = {
  newest: "newest",
  rating: "rating",
  completedJobs: "completedJobs",
  rateAsc: "rateAsc",
  rateDesc: "rateDesc",
} as const;

export type ListBookingsParams = {
  workerId?: number;
  clientWallet?: string;
  status?: ListBookingsStatus;
};

export type ListBookingsStatus =
  (typeof ListBookingsStatus)[keyof typeof ListBookingsStatus];

export const ListBookingsStatus = {
  pending: "pending",
  confirmed: "confirmed",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export type GetRecentBookingsParams = {
  limit?: number;
};

export type GetConversationParams = {
  walletA: string;
  walletB: string;
};