# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Web3 Job Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all job categories
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [manual, online]
        - name: industry
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of job categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
...
  /categories/{id}:
    get:
      operationId: getCategory
      tags: [categories]
      summary: Get a category by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Category details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "404":
          description: Not found
...
  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List bookings
      parameters:
        - name: workerId
          in: query
          schema:
            type: integer
        - name: clientWallet
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, completed, cancelled]
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a new booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingBody"
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get a booking by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "404":
          description: Not found
    patch:
      operationId: updateBookingStatus
      tags: [bookings]
      summary: Update booking status or tx hash
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBookingBody"
      responses:
        "200":
          description: Updated booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
...
  /messages:
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a private message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageBody"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /messages/inbox/{wallet}:
    get:
      operationId: getInbox
      tags: [messages]
      summary: Get all conversations for a wallet
      parameters:
        - name: wallet
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Conversation"
  /messages/conversation:
    get:
      operationId: getConversation
      tags: [messages]
      summary: Get messages between two wallets
      parameters:
        - name: walletA
          in: query
          required: true
          schema:
            type: string
        - name: walletB
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Messages in conversation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
  /messages/{id}/read:
    patch:
      operationId: markMessageRead
      tags: [messages]
      summary: Mark a message as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Message marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
components:
  schemas:
...
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [manual, online]
        industry:
          type: string
        payLevel:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - type
        - industry
        - createdAt
...
    Booking:
      type: object
      properties:
        id:
          type: integer
        workerId:
          type: integer
        workerName:
          type: string
        workerWallet:
          type: string
        clientWallet:
          type: string
        categoryName:
          type: string
        description:
          type: string
          nullable: true
        rateAmount:
          type: string
        rateCurrency:
          type: string
        status:
          type: string
          enum: [pending, confirmed, completed, cancelled]
        txHash:
          type: string
          nullable: true
        scheduledAt:
          type: string
          format: date-time
          nullable: true
        clientRating:
          type: integer
          minimum: 1
          maximum: 5
          nullable: true
        clientReview:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - workerId
        - workerName
        - workerWallet
        - clientWallet
        - categoryName
        - rateAmount
        - rateCurrency
        - status
        - createdAt
    CreateBookingBody:
      type: object
      properties:
        workerId:
          type: integer
        clientWallet:
          type: string
        description:
          type: string
          nullable: true
        scheduledAt:
          type: string
          format: date-time
          nullable: true
      required:
        - workerId
        - clientWallet
    UpdateBookingBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, confirmed, completed, cancelled]
        txHash:
          type: string
          nullable: true
        clientRating:
          type: integer
          minimum: 1
          maximum: 5
          nullable: true
        clientReview:
          type: string
          nullable: true
...
    Message:
      type: object
      properties:
        id:
          type: integer
        fromWallet:
          type: string
        toWallet:
          type: string
        content:
          type: string
        isRead:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - fromWallet
        - toWallet
        - content
        - isRead
        - createdAt
    Conversation:
      type: object
      properties:
        otherWallet:
          type: string
        lastMessage:
          type: string
        lastMessageAt:
          type: string
          format: date-time
        unreadCount:
          type: integer
      required:
        - otherWallet
        - lastMessage
        - lastMessageAt
        - unreadCount
    SendMessageBody:
      type: object
      properties:
        fromWallet:
          type: string
        toWallet:
          type: string
        content:
          type: string
      required:
        - fromWallet
        - toWallet
        - content